week5/my-taxi-pipeline/pipeline/assets/ingestion/requirements.txt
# TODO: Add Python dependencies required by the python ingestion asset.
# - Keep this intentionally minimal.
# - Pin versions for reproducibility.
# Docs: https://getbruin.com/docs/bruin/assets/python
#
pandas==2.2.2
requests==2.32.3
pyarrow==16.1.0
python-dateutil==2.9.0.post0
